move Selector to database crate
