+++ site/src/request_handlers/self_profile.rs
@@ -4,6 +4,7 @@ use std::sync::Arc;
 use std::time::Instant;
 
 use bytes::Buf;
+use database::selector;
 use database::{comparison::Metric, CommitType};
 use headers::{ContentType, Header};
 use hyper::StatusCode;
@@ -12,7 +13,6 @@ use crate::api::self_profile::ArtifactSizeDelta;
 use crate::api::{self_profile, self_profile_processed, self_profile_raw, ServerResult};
 use crate::db::ArtifactId;
 use crate::load::SiteCtxt;
-use crate::selector::{self};
 use crate::self_profile::{get_or_download_self_profile, get_self_profile_raw_data};
 use crate::server::{Response, ResponseHeaders};
